feat(w3.1c): delete Docker sandbox execution layer (path A)
W3.1c removes the Docker-based tool execution path now that W3.1b
switched ShellTool to OsExecutor. Per path A (双层渐进):

DELETED (Docker sandbox **execution** layer — superseded by OsExecutor):
  - sandbox/manager.rs       (697 LOC, SandboxManager + bollard)
  - sandbox/container.rs     (635 LOC, ContainerRunner)
  - sandbox/agent_executor.rs (281 LOC, dead x_claw_agent adapter)
  - tests/agent_hooks_integration_test.rs (314 LOC, depended on SandboxManager)

PRESERVED (out of scope for W3.1c):
  - sandbox/proxy/*          (network allowlist + credential injection by
                              domain — addressed in W3.2)
  - orchestrator/{job_manager,reaper}.rs (background job container layer
                              — separate from tool execution)
  - sandbox/detect.rs        (Docker daemon probe used by setup wizard,
                              boot screen, doctor — not by tool exec)
  - bollard dep              (still needed by orchestrator/{job_manager,reaper})

EXTRACTED:
  - sandbox/docker_conn.rs   (new, 115 LOC) — connect_docker() helper
                              extracted from container.rs so orchestrator
                              and detect.rs share the socket-discovery
                              routine

INLINED:
  - sandbox/os_executor.rs now owns ExecOutput (was super::manager::ExecOutput)

Net diff: -2026 / +315 lines, 4 files removed.

Verification:
  - cargo build -p ironclaw --lib: 0 errors / 0 warnings
  - cargo nextest run --lib sandbox::: 68/68 passed
  - cargo nextest run --lib tools::builtin::shell: 33/33 passed
  - cargo build --tests: clean
  - cargo fmt --all: clean
  - check_no_panics.py: OK

Background reference:
  docs/plans/architecture-refactor/41-docker-vs-os-sandbox-capability-comparison.md
